Extracts common function to higher order function to reduce duplication
